add symble table
can evaluate expression including variable. if the expression is just a
number (such as address, variable, register), it will print hex format.
but this leads to a shift/reduce conflict in parser.y.
also remove some bug in alu.h of dealing with cpu.sf and cpu.of. it can
hit good trap in if-else.c, bubble-sort.c, select-sort.c. but it still
need to include some logic instructions for quick-sort.c.